Fix broken Klayout DRC Report
Loading the drc.klayout.xml in recent KLayout results in an error: 

```
XML parser error: met4.2 is not a valid category path in line 1, column 251
```

Apparently, the category names needs to be surrounded by single quotes. After this fix, KLayout is able to load and show the DRC markers correctly. Verified with KLayout v0.28.10 on Windows 11.