Fix broken FP_PDN_CFG
`PDN_CFG` was renamed to `FP_PDN_CFG` in 2.0.0b7. However, it was not renamed in pdn.tcl, so it is no longer possible to define a custom PDN configuration file.

I verified this fix with the Tiny Tapeout multiplexer: before this fix, hardening the mux controller failed with 2.0.0b7 (where as it worked on 2.0.0b6) since the custom PDN script is not loaded (causing the default PDN to be used, which in turn conflicts with some custom routing we have an causes DRC/LVS failures). 

After applying this fix on top of 2.0.0b7, I was able to harden the mux controller successfully and get similar results to the 2.0.0b6.