Fix regression in `2.1.1` (#548)
* `Odb.SetPowerConnections`

* Fixed an issue introduced in `2.1.1` where modules that are defined as
part of hierarchical netlists would be considered macros and then cause
a crash when they are inevitably not found in the design database. *
Explicitly mention that macros that are not on the top level will not be
connected, and emit warnings if a hierarchical netlist is detected.

## Documentation

* Updated macro documentation to further clarify how instances should be
named and how names should be added to the configuration.